requirements.txt
prometheus-client==0.17.1
slack-sdk==3.21.3
python-dotenv==1.0.0
elasticsearch==8.9.0
requests==2.31.0
pyyaml==6.0.1
jira==3.5.1
schedule==1.2.0
pytest==7.4.0
loguru==0.7.0 